requirements.txt
python-telegram-bot==22.0
python-dotenv==1.1.0
pandas==2.2.3
numpy==2.2.4
ta==0.11.0
requests==2.32.3
flask==2.3.3
